requirements.txt
absl-py==0.8.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cloudant==2.12.0
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
iso8601==0.1.12
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.17.4
oauthlib==3.1.0
opencv-python==3.4.6.27
opt-einsum==3.1.0
paho-mqtt==1.5.0
picamera==1.13
Pillow==6.2.1
pkg-resources==0.0.0
protobuf==3.10.0
psutil==5.6.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
pygame==1.9.6
pymongo==3.9.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
RPi.GPIO==0.7.0
rsa==4.0
six==1.13.0
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wiotp-sdk==0.10.0
wrapt==1.11.2
